Small API interface update (#14)
*Make test_app logic slightly easier.
No need to create two different trakerr clients when the only thing changing is the API key source. Made the if statement simple and easier to understand where the API goes in the sample app.

* Updated test_app
Made if statement easier to look at.

* Updated trakerr.rb
Added parameter typecheck to public CreateAppEvent method.

* Updated trakerr.rb
Added the SendEvent method to send a default error in a more simpler fashion.

* Updated test_app.rb
Added example of SendException and added some comments about the examples.